Link support (#8)
* feat: added support for a Manga's Links `Manga.getLinks()`

* refactor: Update `MustBeDraft` annotation target

The `MustBeDraft` annotation has been updated to only target `ElementType.PARAMETER`. This change ensures that the annotation can only be used on method parameters, providing more clarity and preventing misuse.

* feat: Initialize `demographic` list in MangaAction

The code changes in `MangaAction.java` initialize the `demographic` list with an empty ArrayList. This ensures that the list is not null and can be used without causing NullPointerExceptions.